AGROSMART/requirements.txt
# Web Framework
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5

# HTTP Requests
requests==2.31.0

# WebSocket
python-socketio==5.8.0

# Environment Variables
python-dotenv==1.0.0

# Scheduling
APScheduler==3.10.4

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.9.1
flake8==6.1.0
